Fix rockspec filename inconsistency.
When running "luarocks make", luarocks was complaining about the rockspec filename being inconsistent with the contents.
This was happening because the rockspec file name should be all lowercase. This fix simply renames the file to lowercase in order for luarocks to work.
The package name can be kept as-is as for presentation purposes.